+++ src/components/Breadcrumbs/Breadcrumbs/Breadcrumbs.stories.tsx
@@ -4,12 +4,12 @@ import { Breadcrumbs } from "./Breadcrumbs";
 import { BreadcrumbItem } from "../BreadcrumbItem/BreadcrumbItem";
 
 const meta: Meta<typeof Breadcrumbs> = {
-    component: Breadcrumbs,
-    tags: ["autodocs"],
-    parameters: {
-        docs: {
-            description: {
-                component: "This is a Breadcrumbs wrapper.",
+    component  : Breadcrumbs,
+    tags       : [ "autodocs" ],
+    parameters : {
+        docs : {
+            description : {
+                component : "This is a Breadcrumbs wrapper.",
             },
         },
     },
@@ -19,22 +19,24 @@ export default meta;
 type Story = StoryObj<typeof Breadcrumbs>;
 
 export const Default: Story = {
-    render: (args) => (
+    render : (args) => (
         <Breadcrumbs>
             <BreadcrumbItem>
                 <a href="/">
                     Home
                 </a>
             </BreadcrumbItem>
+
             <BreadcrumbItem>
                 <a href="/category">
-                    Category
+                    Components
                 </a>
             </BreadcrumbItem>
+
             <BreadcrumbItem current>
-                    <span>
-                        Current Page
-                    </span>
+                <a href="/category/current">
+                    Breadcrumbs
+                </a>
             </BreadcrumbItem>
         </Breadcrumbs>
     ),