+++ src/components/Typography/index.tsx
@@ -1,2 +1,2 @@
 export { Text, type TextProps } from "./Text";
-export { Heading, type HeadingProps } from "./Heading";
+export { Heading1, Heading2, Heading3, Heading4, Heading5, Heading6, type HeadingProps } from "./Heading";