+++ src/index.tsx
@@ -143,5 +143,5 @@ export {
 // TYPOGRAPHY ==========================================================================================================
 export {
     Text, type TextProps,
-    Heading, type HeadingProps,
+    Heading1, Heading2, Heading3, Heading4, Heading5, Heading6, type HeadingProps
 } from "./components/Typography";